+++ R/class.R
@@ -153,3 +153,5 @@ setGeneric(
   name = "finalize",
   def = function(x, ...) standardGeneric("finalize")
 )
+
+